Move Makie + CSV deps into package extensions for lean loads
`using GraphEpimodels` pulled in CairoMakie/Makie (and CSV/DataFrames) on
every load — ~19s of Makie precompiled-load time — for code the test suite
never exercises. This made the test suite slow to start.

Move the optional heavy dependencies into package extensions (Julia 1.10+):

- Project.toml: CairoMakie, CSV, DataFrames -> [weakdeps] with two
  [extensions]. Drop Colors (unused) and Makie (only needed transitively
  via CairoMakie).
- ext/GraphEpimodelsCairoMakieExt.jl: all Makie rendering. Visualizer types
  and backend-independent logic (color schemes, frame samplers,
  record_simulation, dispatch) stay in the package, so the public API and
  exports are unchanged; only the drawing method bodies moved.
- ext/GraphEpimodelsPersistenceExt.jl: the two CSV survival-store functions.
  JSON/introspection helpers stay in the package.
- Entry points give an actionable error when their trigger isn't loaded
  (e.g. "save_lattice_plot requires CairoMakie. Run using CairoMakie ...").
- Declare Logging in the test target; the tests use Logging but had been
  relying on it leaking in transitively through the heavy deps.
- Fix examples to load CairoMakie and add an examples/ environment so
  `julia --project=examples ...` works out of the box (Manifest tracked via
  a scoped .gitignore negation).

Result: using GraphEpimodels drops from ~22.9s to ~0.44s (warm); all 1149
tests pass unchanged. Both extensions verified end-to-end (PNG, network
figure, GIF, CSV read/write/append).
