+++ .gitignore
@@ -22,6 +22,9 @@ docs/site/
 # committed for packages, but should be committed for applications that require a static
 # environment.
 Manifest*.toml
+# ...but commit the examples environment's Manifest so `julia --project=examples`
+# resolves the (unregistered, dev-linked) parent package out-of-the-box.
+!examples/Manifest.toml
 
 # File generated by the Preferences package to store local preferences
 LocalPreferences.toml
