+++ Project.toml
@@ -4,21 +4,33 @@ authors = ["Samuel Modée", "Stein Andreas Bethuelsen", "Erik Broman"]
 version = "0.1.0"
 
 [deps]
-CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
-CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
-Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
-DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
-Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 NetworkLayout = "46757867-2c16-5918-afeb-47bfcb05e46a"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
+[weakdeps]
+CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
+CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
+DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
+
+[extensions]
+# Plotting/animation (Makie) loads only when the user runs `using CairoMakie`.
+GraphEpimodelsCairoMakieExt = "CairoMakie"
+# CSV survival-result persistence loads only with `using CSV, DataFrames`.
+GraphEpimodelsPersistenceExt = ["CSV", "DataFrames"]
+
 [compat]
+CSV = "0.10"
+CairoMakie = "0.15"
+DataFrames = "1"
+NetworkLayout = "0.4"
+ProgressMeter = "1"
 julia = "1.10"
 
 [extras]
+Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test"]
+test = ["Logging", "Test"]
