+++ examples/Project.toml
@@ -1,0 +1,4 @@
+[deps]
+CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
+GraphEpimodels = "550e1e89-e91b-4a9b-82ac-1b2c123456ab"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
